// src/app/types/auth.ts
export interface IFormLogin {
  email: string;
  password: string;
}

export interface IFormPlacement {
  placement_type_id: string[];
  ru_title: string;
  kz_title: string;
  en_title: string;
  ru_description: string;
  kz_description: string;
  en_description: string;
  ru_address: string;
  kz_address: string;
  en_address: string;
  lat_lon: string;
  rating: string[];
  region: string[];
  city_id: string[];
  email: string;
  phone: string;
  foods: string[];
  services: string[];
  gallery_images: string[];
}